feat: adapt drop zone areas
