+++ apps/frontend/src/content.ts
@@ -791,7 +791,7 @@ export const Content = {
 	"fileUpload.uploadButton.imgAlt": "Hochladen-Icon",
 	"fileUpload.upload": "Hochladen",
 	"fileUpload.cancel": "Abbrechen",
-	"fileUpload.dropZone.label": "Dateien ablegen, um sie hochzuladen in",
+	"chat.dropZone.label": "Dateien ablegen, um sie zum Chat hinzuzufügen.",
 	//fileUploadButtonStatus
 	"fileUploadButtonStatus.uploading": "Hochladen läuft",
 	"fileUploadButtonStatus.uploaded": "Hochladen abgeschlossen",
